% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_functions.R
\name{break_frame}
\alias{break_frame}
\title{Build a breakpoint dataframe, which describes the line segments of a linear piece-wise
time series function}
\usage{
break_frame(nodes, slopes, initial = 0, slopeseries = NULL)
}
\arguments{
\item{nodes}{a chronological vector of time points at which breakpoint nodes occur,
including the start and end of the full time series.}

\item{slopes}{a vector of slopes for the line segments between each node.
There needs to be one fewer slope than nodes.}

\item{initial}{the initial numeric value of the time series data at the first time
step (i.e., the first value in \code{nodes})}

\item{slopeseries}{an optional time series of linear slopes. if \code{slopeseries}
is not \code{NULL} (the default), then the arguments \code{nodes} and \code{slopes}
are ignored entirely.}
}
\value{
a dataframe object of class \code{break_df}, with the following
columns describing each line segment in the piece-wise time series function.

\describe{
\item{segment}{a numeric identifier of the line segment}
\item{left}{the time at which the left node of the segment occurs}
\item{right}{the time at which the right node of the segment occurs}
\item{start}{the value at the left node of the segment}
\item{end}{the value at the right node of the segment}
\item{slope}{the slope of the segment}
}
}
\description{
Build a breakpoint dataframe, which describes the line segments of a linear piece-wise
time series function
}
\examples{
break_frame(nodes = c(0, 5, 9), slopes = c(.2, -.2))
break_frame(c(0, 5, 9), c(0.2, -.2), initial = 1)
break_frame(c(1, 6, 10), c(.2, -.2), 1)
break_frame(c(0, .2, .5, 1), c(1, 0, -1))
## using a time series of slopes
break_frame(slopeseries = c(rep(.5, 2), rep(0, 2), rep(-.5, 2)))
}
