% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_functions.R
\name{breakwise}
\alias{breakwise}
\title{Build a linear piece-wise time series function from a breakpoint dataframe}
\usage{
breakwise(t, break.df)
}
\arguments{
\item{t}{a vector of time points at which to evaluate the function}

\item{break.df}{a breakpoint dataframe, from which to build the function}
}
\value{
a vector of values, evaluated from the piece-wise function.
}
\description{
Build a linear piece-wise time series function from a breakpoint dataframe
}
\examples{
df = break_frame(nodes = c(0, 5, 9), slopes = c(.2, -.2), 1)
(bw.t = breakwise(0:9, df))
plot(0:9, bw.t, type = "l", xlab = "t", ylab = "breakwise(t)")

# evaluate at non-integer
breakwise(1.32, df)
}
