% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_fit_functions.R
\name{example_fitfun}
\alias{example_fitfun}
\title{Example of a fit_FUN, passed to simulate_spacebreak}
\usage{
example_fitfun(Y, ...)
}
\arguments{
\item{Y}{a matrix whose rows contain pixels and whose columns contain
timepoints.}

\item{...}{additional arguments}
}
\description{
Example of a fit_FUN, passed to simulate_spacebreak
}
