% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{generate_spcov}
\alias{generate_spcov}
\title{Generate pixels with spatial correlations for a grid-based map}
\usage{
generate_spcov(
  map.dims = c(5, 5),
  unit.map = TRUE,
  covar_FUN = remotePARTS::covar_exp,
  covar.pars = list(range = 1),
  sd = 1
)
}
\arguments{
\item{map.dims}{a dimension vector of length 2, containing the rows and
columns, respectively of the map. \code{map.dims} are coerced to integers
with \code{as.integer(map.dims)}. The number of pixels (n) in the resulting
map is equal to \code{map.dims[1]*map.dims[2]}.}

\item{unit.map}{logical: should a unit map be used (range from 0 to 1)?}

\item{covar_FUN}{function to generate distance-based correlations from.
\code{cavar_FUN} should accept an nxn distance matrix (\code{D.mat}) and
return a nxn covariance matrix.}

\item{covar.pars}{a list of parameters used by \code{covar_FUN}.}

\item{sd}{the standard deviation of the system. The covariance matrix
that results from \code{covar_FUN(D.mat, ...)} is scaled by
\code{sqrt(sd)}.}
}
\value{
a list with the following elements

\describe{
\item{coord.df}{a data.frame with n rows and columns corresponding
to pixel identifiers (\code{pixel}), the pixel's x-coordinate
(\code{coords.x}) and y-coordinate (\code{coords.y})}
\item{D.mat}{the nxn distance matrix, containing pairwise distances
among all pixels}
\item{covar.mat}{the nxn covariance matrix, containing pairwise
covariances among all pixels}

}
}
\description{
Generate pixels with spatial correlations for a grid-based map
}
\examples{
results = generate_spcov(c(5, 5))
results$coord.df
results$covar.mat
results$D.mat

## smaller sd
generate_spcov(c(5, 5), sd = .5)$covar.mat

## unequal map dimensions
generate_spcov(c(2, 5))$coord.df

## plot distance-based covariance function
plot(results$covar.mat ~ results$D.mat, xlab = "distance",
     ylab = "covariance")
}
