% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_functions.R
\name{get_breaks}
\alias{get_breaks}
\title{Get breakpoints and nodes from a time series of slopes}
\usage{
get_breaks(slope.vec)
}
\arguments{
\item{slope.vec}{a vector of linear slopes for each time point of a
time series}
}
\value{
a list containing vectors of \code{nodes} and \code{slopes}
for each line segment
}
\description{
Get breakpoints and nodes from a time series of slopes
}
\examples{
(slope_vector = c(rep(0, 3), rep(.2, 4), rep(-.1, 1)))
(breakR:::get_breaks(slope.vec = slope_vector))
}
