% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_fit_functions.R
\name{melt_spacebreak}
\alias{melt_spacebreak}
\title{Melt a spatiotemporal matrix into a long-form data frame}
\usage{
melt_spacebreak(Y)
}
\arguments{
\item{Y}{a spatiotemporal matrix, as results from \code{generate_spacebreak()}.}
}
\value{
a data frame with three columns: indicating the \code{pixel} ID,
the \code{time} point, and the data \code{value}.
}
\description{
Melt a spatiotemporal matrix into a long-form data frame
}
