% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-simulation-function.R
\name{plot.spacebreak}
\alias{plot.spacebreak}
\title{plot a spacebreak object}
\usage{
\method{plot}{spacebreak}(x, type = "ts", ...)
}
\arguments{
\item{x}{the spacebreak object}

\item{type}{a character vector specifying which type of plot to make}

\item{...}{additional parameters passed to other plot methods}
}
\description{
plot a spacebreak object
}
