% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_funcitons.R
\name{season}
\alias{season}
\title{Generate seasonality with a sinusoidal function}
\usage{
season(t = 0:24, amp = 1, per = 12, noise.sd = 0, trend = 0)
}
\arguments{
\item{t}{a vector of time points at which to evaluate the sinusoidal function}

\item{amp}{the amplitude of the sine wave}

\item{per}{periodicity of the wave: the number of time points over which one full
cycle is completed}

\item{noise.sd}{the standard deviation of the random normal variable, which
is added to the sinusoidal wave.}

\item{trend}{an optional linear trend coefficient to add to the sine wave.}
}
\value{
a vector of values the same length as \code{t}
}
\description{
Generate seasonality with a sinusoidal function
}
\details{
a sine wave is generated as \code{amp + sin(2*pi / per * t) + trend*t + e} where
\code{e} is a random normal variable with sd = \code{noise.sd}.
}
\examples{
season()
season(amp = .8)
season(t = seq(0, 1, .1), per = 1)
season(t = 15)
season(noise.sd = 1)

# visualize
plot(season(), type = "l", ylim = c(-1, 1))
plot(season(amp = .5), col = "red", type = "l", ylim = c(-1, 1))
plot(season(amp = .5, noise.sd = .5), col = "blue", type = "l")
plot(season(0:100, trend = .1/12), type = "l")
}
