% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemporal_functions.R
\name{sparima_sim}
\alias{sparima_sim}
\title{Simulate an ARIMA model with spatially correlated errors}
\usage{
sparima_sim(
  ntime,
  covar,
  fast = FALSE,
  model = list(ar = 0.2),
  innov = NULL,
  ...
)
}
\arguments{
\item{ntime}{number of time points}

\item{covar}{matrix containing pairwise covariances among pixels}

\item{fast}{logical: should fast method be used? see \code{rnorm_spcor}}

\item{model}{a named list containing model specification for \code{arima.sim()}.
This list should NOT include an 'innov' element.}

\item{innov}{optional matrix of innovations to use in place of \code{ntime},
\code{covar}, and \code{fast}.}

\item{...}{additional arguments passed to \code{arima.sim()}}
}
\value{
a matrix containing the simulated variable. Columns correspond to
time points and rows correspond to locations (i.e., pixels).
}
\description{
Simulate an ARIMA model with spatially correlated errors
}
\examples{
cor.mat <- matrix(c(1.0, 0.3, 0.1,
                    0.3, 1.0, 0.0,
                    0.1, 0.0, 1.0), nrow = 3)

sparima_sim(20, cor.mat)
sparima_sim(innov = rnorm_spcor(20, cor.mat))
## specify burn-in period (from arima.sim)
sparima_sim(20, cor.mat, n.start = 20)
## single time point (after 100)
sparima_sim(1, cor.mat, n.start = 100)

## compare observed covariance to true covariance
spcov = generate_spcov(c(10, 10), sd = 1, unit.map = TRUE)
D = spcov$D.mat
C = spcov$covar.mat
x <- sparima_sim(50, C, model = list(ar = .8), n.start = 100)
## sample covariance vs. distance
plot(x = D, y = cor(t(x)), xlab = "distance", ylab = "covariance")
## add line for covariance function
curve(remotePARTS::covar_exp(x, 1), from = 0, to = max(D),
      add = TRUE, col = "red")
}
